requirements.txt
openai
flask
flask_httpauth
python-dotenv
pydub